+++ MS/mlaas/modeling/utils/model_utils/sklearn_regression/randomforest_regression.py
@@ -50,9 +50,10 @@ class RandomForestRegressionClass:
 
         self.n_estimators = int(hyperparameters['n_estimators'])
         self.criterion = hyperparameters['criterion']
-        self.max_features = int(hyperparameters['max_features'])
-        self.min_impurity_decrease = int(hyperparameters['min_impurity_decrease'])
+        self.max_features = hyperparameters['max_features']
+        self.min_impurity_decrease = float(hyperparameters['min_impurity_decrease'])
         self.min_samples_leaf = int(hyperparameters['min_samples_leaf'])
+        
         if hyperparameters['max_depth'] != 'None':
             self.max_depth = int(hyperparameters['max_depth'])
         else:
